requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4; platform_system == "Windows"
flask==2.0.1
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-dialogflow==2.1.2
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
itsdangerous==2.0.1
jinja2==3.0.1
markupsafe==2.0.1
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2; python_version >= "3.6"
six==1.15.0
slack-bolt==1.5.0
slack-sdk==3.5.0
urllib3==1.26.4
werkzeug==2.0.1
